requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv

pillow
fpdf2
qrcode
cairosvg